feat : added indexing

+++ app/core/config.py
@@ -62,9 +62,9 @@ class Settings(BaseSettings):
     FRONTEND_BASE_URL: str = "http://localhost:5173"
 
     # Form timing (reminders, escalations, expiry)
-    FORM_REMINDER_HOURS: int = 2
-    FORM_ESCALATION_HOURS: int = 3
-    FORM_EXPIRY_HOURS: int = 24
+    FORM_REMINDER_HOURS: float = 2
+    FORM_ESCALATION_HOURS: float = 3
+    FORM_EXPIRY_HOURS: float = 24
 
     model_config = {"env_file": ".env", "env_file_encoding": "utf-8", "case_sensitive": True}
